requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<0.30
httpx>=0.27,<1.0
redis>=5.0,<6.0
pydantic>=2.4,<3.0
pydantic-settings>=2.0,<3.0
python-dotenv>=1.0,<2.0
google-genai>=0.1.0
python-json-logger>=2.0.7,<3.0
